+++ jdaviz/core/loaders/resolvers/file/file.vue
@@ -19,6 +19,10 @@
     :api_hints_enabled="api_hints_enabled"
     :valid_import_formats="valid_import_formats"
     :server_is_remote="server_is_remote"
+    :is_wcs_linked="is_wcs_linked"
+    :image_data_loaded="image_data_loaded"
+    :footprint_select_icon="footprint_select_icon"
+    :custom_toolbar_enabled="custom_toolbar_enabled"
   >
     <v-row style="padding-left: 12px; margin-bottom: 16px">
       Select a file with data you want to load into this instance of Jdaviz.